cranky: Get rid of helper-wrapper
Now that 'cranky' is the main entry point into the toolset, we can add
the functionality of 'helper-wrapper' to 'cranky' and get rid of the
wrapper.

Figure out if the provided command is a helper command and if so, do
what 'helper-wrapper' used to do. Otherwise simply exec the provided
command.

Signed-off-by: Juerg Haefliger <juergh@canonical.com>